Used a makefile, resolved error's

+++ main.c
@@ -1,11 +1,10 @@
 #include <stdio.h>
 #include <math.h>
-#include "functional.c"
 
 int main () {
-    float a, b, c;
+    int a, b, c;
     printf( "Input coefficient's of equation (a, b, c)\n" );
-    scanf( "%f %f %f", &a, &b, &c );
+    scanf( "%d %d %d", &a, &b, &c );
     process( a, b, c );
     return 0;
 }
